requirements.txt
numpy==1.24.3
matplotlib==3.7.2
gymnasium==0.29.1
pyyaml==6.0.1
tqdm==4.65.0
